Linearity correction - TMC codes for calibration object, variable stored as uint8, range from 1.03 to 1.20

+++ Firmware/Dcodes.cpp
@@ -465,6 +465,7 @@ void dcode_10()
 void dcode_12()
 {//Time
 	LOG("D12 - Time\n");
+
 }
 
 
@@ -590,7 +591,7 @@ void dcode_2130()
 			}
 			else if (strncmp(strchr_pointer + 7, "wave", 4) == 0)
 			{
-				uint16_t fac1000 = atoi(strchr_pointer + 11) & 0xffff;
+				uint8_t fac1000 = atoi(strchr_pointer + 11) & 0xffff;
 				if (fac1000 < TMC2130_WAVE_FAC1000_MIN) fac1000 = 0;
 				if (fac1000 > TMC2130_WAVE_FAC1000_MAX) fac1000 = TMC2130_WAVE_FAC1000_MAX;
 				tmc2130_set_wave(axis, 247, fac1000);
